Add files via upload
Added in George Socks as a character. 

TODO: Will be a future romance route. I plan on adding in Noah Caldwell Gervais and Turbo Button as characters (though the latter won't be a romance route). After that, I plan on expanding upon existing routes before adding in more romance options.